+++ backend/src/common/decorators/roles.decorator.ts
@@ -1,7 +1,0 @@
-import { SetMetadata } from "@nestjs/common";
-
-export const ROLES_KEY = "roles";
-
-/** Marks a route as requiring one of the given role keys (see RolesGuard). */
-export const Roles = (...roles: string[]): MethodDecorator & ClassDecorator =>
-  SetMetadata(ROLES_KEY, roles);